#308 split SchemaPolicy subject list logic to DefLogic
